// packages/oauth-client/src/types.ts
export interface ClientConfig {
  readonly clientId: string;
  readonly clientSecret: string;
  readonly redirectUri: string;
  readonly scope: string;
}

export interface AdapterConfig {
  readonly authorizationUri: string;
  readonly tokenUri: string;
  readonly profileUri: string;
}
// ...
export interface ErrorResponse {
  /**
   * A single error code.
   * REQUIRED.
   */
  readonly error: ErrorCode;
  /**
   * Human-readable text providing additional information,
   * used to assist the client developer in understanding
   * the error that occurred.
   * OPTIONAL.
   */
  readonly error_description?: string;
  /**
   * A URI identifying a human-readable web page with information
   * about the error, used to provide the client developer
   * with additional information about the error.
   * OPTIONAL.
   */
  readonly error_uri?: string;
}
// ...
export type ErrorCode =
  | "invalid_request"
  | "invalid_client"
  | "invalid_grant"
  | "invalid_scope"
  | "unauthorized_client"
  | "unsupported_grant_type";